Fixed #293
Fix the case by doing something completely different than is called for in the bug report.

We following the layout of what the V3 HTML document says to do.  This means that we use the child text of the xref when it exists to the exclusion of any generated text.  When the child text does not exist then we use the synthesized text string as the text for the anchor element.  In all cases the anchor element is emitted with an href of the target.

Uses of the "cite" node are removed as they did not seem to be useful.

In a step forward we emit a class on the a node of "xref" per the V3 HTML document.